Full Site Editing: remove obsolete wp_template CPT (#35075)
With the new themes approach we don't really use wp_template CPT,
and unnecessarily complicates a lot of our code. I'm also removing some
functionality that was tied to it like sidebar selector and separate handling 
in blocks for this page. All of this will be unreachable once we hide the
template CPTs.

This separation made sense when we were going for the Blank theme
approach, or full page hijacking in the plugin. Now this layout is defined
by theme template files (header.php, footer.php) so having this stored
in the template becomes kind of pointless since it can't be manipulated.
Even if we do it, it wouldn't be reflected on the page.

Follow up to this will be to rename wp_template_part to wp_template.