Checkout: Disable stripe fields while loading stripe (#35196)
* lib/stripe: Add withStripe HOC

* lib/stripe: Pass stripeConfiguration to WrappedComponent

* lib/stripe: Document resolved value of createStripePaymentMethod

* lib/stripe: Allow any args passed to getStripeConfiguration

Also `useStripeConfiguration` and `withStripe` since both of those call
that function.

* lib/stripe: add isStripeLoading to withStripe HOC

* StripeElementsPaymentBox: Use withStripe HOC

* Move CreditCardFormFields styles to the component

* CreditCardFormFields: Disable fields on isStripeLoading

Note that if `isStripeLoading` is true, `stripe` will always be
undefined, so we can't even render the Stripe Elements fields. Therefore
when `isStripeLoading` is true we render and disable the non-stripe form
fields. We then ignore `isStripeLoading` when `stripe` is defined
because it will always be false.

* Checkout: Pass isStripeLoading through to CreditCardFormFields

* lib/stripe: Disable isStripeLoading before window.Stripe

In case `window.Stripe()` has an error, this will make sure that the
old-style form fields are used as a fallback.